// src/modules/reports/dto/report-interfaces.ts
export interface ITabularOrderBy {
  draggedId: string;
  columnDisplayName: string;
  orderBy: string;
}
// ...
export interface IOperationColumn {
  isOperation: boolean;
  isCustomColumn: boolean;
  draggedId: string;
}

export interface IOperationOperatorColumn {
  isOperation: boolean;
  operator: string;
}

export type OperationColumnType = IOperationOperatorColumn | IOperationColumn;
// ...
export interface ICustomOperationColumn {
  draggedId: string;
  columnDisplayName: string;
  footerAggregation: Array<string>;
  type: string;
  operation: string;
  savedTokens: Array<OperationColumnType>;
  customColumn: boolean;
  customColumnType: string;
  hidden: boolean;
  pinned: boolean;
  index: number;
  trunc: boolean;
  round: boolean;
  trValue: number;
  selected?: boolean;
  isMetric?: boolean;
  isExplodedBy?: boolean;
}
// ...
export interface IChartData {
  id: string;
  name: string;
  type: string;
  orderIndex: number;
  [key: string]: unknown;
}

// --- Tabular Header ---

export interface ITabularHeader {
  text: string;
  datafield: string;
  columnName?: string;
  aggregates: string[];
  pinned: boolean;
  hidden: boolean;
  draggedId?: string;
  editable?: boolean;
  columntype?: string;
  index?: number;
  headerColumnType?: string;
}